31: move model.vue into new folder
